chore: Move photo gallery layout into lib
