+++ src/routes/events/2017/photos/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import PhotoLayout from '../../PhotoLayout.svelte';
+	import PhotoLayout from '../../../../lib/layout/PhotoLayout.svelte';
 	import type { PageData } from './$types';
 
 	export let data: PageData;
